requirements.txt
matplotlib
Pillow
torch
textwrap